New processing methods and complete Wikidata item creation notebook
I reworked a number of processes in the ecoregion data processing workflow after reviewing through the prepped data and finding some mistakes. This changed a bunch of things in the underlying functions and data processing sequence. I also worked up a full example for creating items using a fork of wikidataintegrator and captured notes on the process.